Added setup for Student CRUD application with Django. - Added initial Django project structure with student_app for managing students, faculty, subjects, and marks. - Implemented models for User, Student, Faculty, Subject, and Mark with appropriate relationships. - Registered models in Django admin - Created serializers for user registration and login, including handling of student and faculty roles. - Developed API views for user registration, login, and CRUD operations on subjects and marks. - Configured Django settings for REST framework and JWT authentication. - Added URL routing for API endpoints.

+++ practice django/vege/views.py
@@ -1,9 +1,12 @@
+from multiprocessing import context
 from django.shortcuts import render, redirect
 from .models import *
 from django.http import HttpResponse
+from rest_framework.decorators import api_view # type: ignore
 # Create your views here.
 
-def receipes(request):
+@api_view(['POST'])
+def  create_receipes(request):
 
     if request.method == "POST":
 
@@ -21,16 +24,29 @@ def receipes(request):
             receipe_name=receipe_name,
             receipe_description=receipe_description)
 
-        return redirect('/receipes/')
-    
-    queryset = Receipe.objects.all()
-    context = {'receipes':queryset}
+        # return redirect('/receipes/'), 
+        return HttpResponse({'status':'success','message':'data created successfully ',"data":context}, status=201)
+
+
+    # queryset = Receipe.objects.all()
+    # context = {'receipes':queryset}
     return render(request, 'receipes.html',context)
+    # return HttpResponse({'status':'success','message':'data fetched successfully ',"data":context}, status=200) 
 
-def update_receipe(request,id):
-    queryset = Receipe.objects.get(id =id)
+@api_view(['GET'])
+def receipe_detail(request):
+    id = request.GET.get('id')
+    queryset = Receipe.objects.get(id=id)
+    context = {'receipe': queryset}
+    # return render(request, 'receipe_detail.html', context), 
+    return HttpResponse({'status':'success','message':'data fetched successfully ',"data":context}, status=200)
 
-    if request.method == "POST":
+@api_view(['PUT', 'POST'])
+def update_receipe(request):
+    id = request.GET.get('id')
+    queryset = Receipe.objects.get(id=id)
+
+    if request.method == "PUT" or request.method == "POST":
         data = request.POST
 
         receipe_image = request.FILES.get('receipe_image')
@@ -43,13 +59,18 @@ def update_receipe(request,id):
         if receipe_image:
             queryset.receipe_image = receipe_image
         queryset.save()
-        return redirect('/receipes/')
+        #return redirect('/receipes/'), 
+        return HttpResponse({'status':'success','message':'data updated successfully ',"data":context}, status=200)
 
     context = {'receipe': queryset}
 
     return render(request, 'update_receipes.html', context)
 
-def delete_receipes(request,id):
-    queryset = Receipe.objects.get(id = id)
-    queryset.delete()
-    return redirect('/receipes/')
+@api_view(['DELETE'])
+def delete_receipes(request):
+    if request.method == "DELETE":
+        id = request.GET.get('id')
+        queryset = Receipe.objects.get(id=id)
+        queryset.delete()
+        #return redirect('/receipes/'), 
+        return HttpResponse({'status':'success','message':'data deleted successfully ',"data":context}, status=200)
